Add GRN re-use capability

+++ src/preprocessing/grn_creation.py
@@ -7,6 +7,7 @@ import scanpy as sc
 from arboreto.algo import grnboost2
 from tabulate import tabulate
 from scipy import sparse
+from pathlib import Path
 
 
 def create_GRN(cfg: ConfigParser) -> None:
@@ -31,13 +32,16 @@ def create_GRN(cfg: ConfigParser) -> None:
         real_cells.X = real_cells.X.todense()
 
     # preparing GRNBoost2's input
-    real_cells_df = pd.DataFrame(real_cells.X, columns=real_cells.var_names)
+    if not Path(cfg.get("GRN Preparation", "Inferred GRN")).exists():
+        real_cells_df = pd.DataFrame(real_cells.X, columns=real_cells.var_names)
 
-    # we can optionally pass a list of TFs to GRNBoost2
-    print(f"Using {len(TFs)} TFs for GRN inference.")
-    real_grn = grnboost2(real_cells_df, tf_names=TFs, verbose=True, seed=1)
-    real_grn.to_csv(cfg.get("GRN Preparation", "Inferred GRN"))
-    print("Successfully saved GRN inferred by GRNBoost2 GRN to", cfg.get("GRN Preparation", "Inferred GRN"))
+        # we can optionally pass a list of TFs to GRNBoost2
+        print(f"Using {len(TFs)} TFs for GRN inference.")
+        real_grn = grnboost2(real_cells_df, tf_names=TFs, verbose=True, seed=1)
+        real_grn.to_csv(cfg.get("GRN Preparation", "Inferred GRN"))
+        print("Successfully saved GRN inferred by GRNBoost2 GRN to", cfg.get("GRN Preparation", "Inferred GRN"))
+    else:
+        print("Using already existing GRNBoost2 GRN at", cfg.get("GRN Preparation", "Inferred GRN"))
 
     # read GRN csv output, group TFs regulating genes, sort by importance
     real_grn = (
